+++ ijkmedia/ijkplayer/ff_ffplay_options.h
@@ -153,6 +153,8 @@ static const AVOption ffp_context_options[] = {
         OPTION_OFFSET(enable_accurate_seek),       OPTION_INT(0, 0, 1) },
     { "skip-calc-frame-rate",                      "don't calculate real frame rate",
         OPTION_OFFSET(skip_calc_frame_rate),       OPTION_INT(0, 0, 1) },
+    { "get-frame-mode",                      "warning, this option only for get frame",
+        OPTION_OFFSET(get_frame_mode),       OPTION_INT(0, 0, 1) },
 
         // iOS only options
     { "videotoolbox",                       "VideoToolbox: enable",