+++ ijkmedia/ijkplayer/ijkavformat/ijkioprotocol.c
@@ -27,6 +27,7 @@ extern IjkURLProtocol ijkio_ffio_protocol;
 extern IjkURLProtocol ijkio_androidio_protocol;
 #endif
 extern IjkURLProtocol ijkio_cache_protocol;
+extern IjkURLProtocol ijkio_httphook_protocol;
 
 int ijkio_alloc_url(IjkURLContext **ph, const char *url) {
     if (!ph) {
@@ -42,6 +43,10 @@ int ijkio_alloc_url(IjkURLContext **ph, const char *url) {
         h = (IjkURLContext *)calloc(1, sizeof(IjkURLContext));
         h->prot = &ijkio_ffio_protocol;
         h->priv_data = calloc(1, ijkio_ffio_protocol.priv_data_size);
+    } else if (!strncmp(url, "httphook:", strlen("httphook:"))) {
+        h = (IjkURLContext *)calloc(1, sizeof(IjkURLContext));
+        h->prot = &ijkio_httphook_protocol;
+        h->priv_data = calloc(1, ijkio_httphook_protocol.priv_data_size);
     }
 #ifdef __ANDROID__
       else if (!strncmp(url, "androidio:", strlen("androidio:"))) {